revision based on pr feedback, critical requirement oversight was that legacy metrics was never released and thus does not need continued support, it can be replaced wholesale.
