+++ CHANGELOG.md
@@ -6,17 +6,18 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Canonical metrics mode: opt-in harmonized metric surface via `WORKER_CANONICAL_METRICS=true` — [details](conductor-client-metrics/README.md#detailed-technical-notes--unreleased)
-- Automatic metrics wiring: `ConductorClient.Builder.withMetricsCollector(...)` installs the HTTP interceptor and auto-registers listeners on `TaskClient` and `WorkflowClient` (automatic in canonical mode; opt-in via `setAutoWiringEnabled(true)` for legacy)
+- Standardized Prometheus metrics: `PrometheusMetricsCollector` now emits the harmonized cross-SDK metric surface — [details](conductor-client-metrics/README.md)
+- Automatic metrics wiring: `ConductorClient.Builder.withMetricsCollector(...)` installs the HTTP interceptor and auto-registers listeners on `TaskClient`, `WorkflowClient`, and `TaskRunnerConfigurer`
+- HTTP API client metrics via OkHttp interceptor (`http_api_client_request_seconds`, `task_result_size_bytes`, `workflow_input_size_bytes`)
+- Event-driven metrics architecture with `EventDispatcher` and typed event POJOs
 
 ### Changed
 
-- Legacy metrics emit unchanged by default; no env var required
+- `PrometheusMetricsCollector` metric names updated to the harmonized cross-SDK catalog (e.g. `task_poll_total`, `task_execute_time_seconds`)
 - `micrometer-registry-prometheus` is now a transitive (`api`) dependency
 
 ### Deprecated
 
-- `PrometheusMetricsCollector` — use `MetricsCollectorFactory.create()` or `MetricsBundle.create()`
 - `TaskClient.ack(String, String)` — use `ack(String taskType, String taskId, String workerId)`
 
 ## [4.0.0] - 2024-10-09
