+++ gradle.properties
@@ -1,1 +1,1 @@
-version=4.2.0
+version=5.1.0
